+++ frontend/src/components/main/ticketsPageTemp/index.css
