+++ frontend/src/components/main/ticketsPageTemp/ticketScroller/index.css
