+++ frontend/src/components/main/ticketsPageTemp/ticketTile/index.css
