+++ frontend/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
-import TicketTrack from './components/tickettrack';
+import TicketTrack from './components/tickettrackTemp';
 import './index.css';
 
 const rootElement = document.getElementById('root');
